% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_0s_col.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{to remove columns that contain only 0s}
\usage{
rm_0s_by_col(input_matrix)
}
\arguments{
\item{input_matrix, }{must be matrix}
}
\value{
, a matrix with no columns that only contain 0
}
\description{
to remove columns that contain only 0s
}
