% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_0s_row.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{to remove rows that contain only 0s}
\usage{
rm_0s_by_row(input_matrix)
}
\arguments{
\item{input_matrix, }{must be matrix}
}
\value{
a matrix, with no rows that only contain 0
}
\description{
to remove rows that contain only 0s
}
